+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* Fixed blacklist issue
+* Fixed bug in Discord message not being posted
+* Added version check at startup
+* Added tests for the filters